requirements.txt
paddlepaddle>=2.6.0
paddleocr>=2.7.0
openai>=1.0.0
pyyaml
python-dotenv
opencv-python-headless